python: improve ergonomics of watcher callbacks
Problem: The Python versions of various watcher callbacks are
inflexible in that they only allow a single "args" parameter. This not
only means that the "args" parameter is required for callbacks that
don't need any extra args, but for callbacks that do need extra args,
the arguments must be packed up into a single args object.

Improve this situation by allowing callbacks to take arbitrary args and
keyword arguments. These are then passed along as-is to the specified
callback, allowing the callbacks to take either no arguments, or an
arbitrary list of arguments and keyword arguments.

Unfortunately, this requires establishment of a compatibility mode
to support the previous mode with a default, required `args` keyword
parameter with default `args=None`. The compat mode is activated only
if the callback in question expects the correct number of arguments,
and there were no unnamed arguments (args is None) passed in the
constructor.

This compat mode can be removed in the future when there are no extant
users of the old callback mode.