job-ingest: add disable-validator module option
Problem: There is no way to load the job-ingset module and disable
job validation, but this could be useful as an alternative to loading
a "permissive" validator, when benchmarking, or when running a high-
throughput workload where all submitted jobs are known to be valid.

Add a new job-ingest module option `disable-validator` which disables
the job validator completely.