flux-mini: add --timing option to flux-mini alloc --bg
Problem: There is no way to easily capture the startup timing of a Flux
instance.

Add a --timing option for testing to the flux-mini alloc subcommand.
When used with the --bg option, this will print a set of timings to
stdout in the form (with optional --timing-header, header row is also
printed)

NODES  SIZE FNOUT    T_ALLOC      T_URI    T_READY    (TOTAL) T_SHUTDOWN
    4   256     2     2.6636     0.9883    21.2224    24.8744     0.0029