job-exec: make jobinfo_set_expiration() idempotent
Problem: The expiration time of a running job might be updated, but
the jobinfo_set_expiration() function only works when first starting
the job, so it can't be used for the update.

Make the jobinfo_set_expiration() call idempotent. That is, allow the
call to be made at any point while a job is running. Any existing
timer will be destroyed, and a new timer started for the current
expiration time obtained from R.