job-exec: watch for R updates
Problem: For jobs that have an expiration, the job-exec module
schedules a timer watcher to fire when the expiration elapses so
that it can raise a timeout exception on the job. RFC 21 defines a
resource-update event which may adjust the job expiration, but the
job-exec module currently has no way of being notified of the update
so it can reschedule its timer.

Watch for R updates using the new job-info.update-watch RPC. On any
update, destroy and re-add the expiration timer so that it is always
up to date.