job-manager: support duration updates for running jobs
Problem: The job manager update service does not currently support
updates to running jobs, but updates of the expiration of running
jobs may be allowed by the update-duration plugin.

For running jobs with a duration update, issue a RFC 27
sched.expiration RPC to notify the scheduler of the update request. If
the RPC is successful or returns an ENOSYS error, then the update
is allowed. Additionally issue a resource-update event to modify
expiration of a running job.

Add a post_resource_updates() function to handle posting
resource-update events for running jobs when necessary. At this time,
only an update of duration for a running ojb is supported.