resource: monitor assigned resources for expiration update
Problem: The resource module fetches R fom the parent job-info module
when the instance is being run as a job (method=job-info), but it
only gets a copy of R at startup and therefore ignores any updates
to the resource set from resource-update events posted to the job
eventlog.

Switch from the job-info.lookup RPC to job-info.update-watch so that
the resource module receives any updates to R. Currently, assume that
the only supported update is an adjustment to the expiration. Process
the job expiration update and post a `resource-update` event to the
resource.eventlog containing the new expiration.