job-manager: update running jobs if necessary after R update
Problem: Jobs that do not specify an explicit duration may have their
expiration automatically set to the expiration of the current resource
set as seen by the scheduler. If an expiration update is then issued
for the enclosing instance, these running jobs will be terminated
before the new instance expiration.

Add a watch of resource.R to the job manager update service. If an
update occurs that modifies the resource set expiration, then walk the
list of running jobs and update any job's expiration if it exactly
matches the old expiration, since this indicates the expiration was
automatically set be the scheduler.