resource: improve errors from resource.*status RPCs
Problem: When a resource.status or resource.sched-status RPC is send
to the resource module and inventory->R is not yet set, tools such as
`flux resource list` report the mysterious error:

 flux-resource: ERROR: ENOENT: No such file or directory

For this specific case, set a human readable error string to

 resource inventory currently unavailable

To give a better idea of what's gone wrong.

Improve the error messages for several other (very unlikely) cases
in the response handling.

Fixes #5882