testsuite: add exit-on-error to job in t4583-free-range-test.sh
Problem: The job on which a node fails may not be terminated
immediately because the exception raised is non fatal. This can make
the test in issues/t4583-free-range-test.sh take longer than necessary
or fail becaue the second 3 node job is waiting for resources to
be free.

Add `-o exit-on-error` to the previous job so it exits immediately
when the rank is forced down.