cli: add `-B,--bank=` option to flux submission commands
Problem: Users are forced to use --setattr=bank=BANK to set a bank
attribute in their jobspec, but this is an attribute documented
in RFC 14 and is commonly required in production system instance
usage.

Add a `-B,--bank=` to all flux submission cli commands to make
use of banks more user friendly.