cmd: add flux-modprobe prototype
Problem: The serial rc scripts are slower than necessary, and do
not use a config file or other means to determine module relationships.
A new tool is warranted.

Add a prototype of a flux-modprobe utility that uses a TOML config
to represent modules, their dependencies, and rc1/rc3 tasks. The
utility then uses this information to runs startup/shutdown commands
and load/unload modules as quickly as possible.