perilog: avoid potential use-after-free
Problem: The perilog_per_rank() relies on accessing the global
perilog_config for determining if a prolog or epilog is configured in
per-rank mode or not, but the config can be updated while a prolog or
epilog for a job is active. In the best case, this returns an invalid
result, but in the worst case the config entry for prolog/epilog
could have been freed which would result in a crash.

Store the per-rank value with the perilog_proc so that we don't
rely on the global config at all when testing the per-rank value for
a prolog/epilog. Drop the perilog_per_rank() function since it is
no longer necessary.