testsuite: decrease racy tests in t2274-manager-perilog.t
Problem: A few tests in t2274-manager-perilog.t are racy. The cancel
test can fail if the cancel is sent before the flux-perilog-run process
is ready, causing the prolog to time out instead of being terminated.
In another test, flux-perilog-run may exit with failure instead
of being signaled. Finally, a timeout is too short which can cause
multiple ranks to be drained instead of the expected single rank.

Remove the cancel test, which has a less racy duplicate in
t2274-manager-perilog-per-rank.t.

Update one case to use test_must_fail_or_be_terminated.

And finally, increase the timeout to 2s to give time for planned
successful ranks to exit before they are signaled by the test.