libsubprocess: workaround potential deadlock in pre-exec hook
Problem: If a pre-exec hook generates a lot of output and fills stdio
buffers, then the subprocess child could block inside the pre-exec
hook. Since the parent is waiting for the sync fd to close, this
could cause a deadlock.

Set the stdout and stderr fds to non-blocking for the duration of
the pre-exec hook, which should turn a potential hang to errors in
the pre-exec hook code instead.