shell: output: write shell.output data to correct task
Problem: Data written via the shell.output plugin topic is currently
always written to the same destination as the first local task.
This leads to data being written to the wrong file when there is a
file open per task.

Use task_output_list_write() to direct data to the correct task.