python: add JournalConsumerBase abstract base class
Problem: The JournalConsumer class in flux.job is specific to the
job manager journal, but there may be other services that
implement a journal-like interface.

Add an abstract base class for journal consumer interfaces in
flux.abc.JournalConsumerBase. Include a JournalEventBase class that
derives from EventLogEvent.

Most of the journal consumer interface is implemented in this class,
with the expectation that derived classes only need to implement the
few abstract methods left open for implementations.