librlist: add rnode_diff_ex()
Problem: It will be necessary to exclude some resources from the
results of rnode_diff() in order to support the exclusion of
certain resources from resource verification, but the rnode class
doesn't have a simple way to do that.

Add rnode_diff_ex() which takes a mask of resources to exclude from
the diff. After computing the diff, zero out ignored resources.

Support only cores and gpus for now.